requirements.txt
django == 5.2
django-extensions == 4.1
djangorestframework == 3.16.0